Allow to process new payment models locally
