+++ src/wstore/charging_engine/views.py
@@ -20,17 +20,13 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 import json
-import requests
 
 from copy import deepcopy
 from logging import getLogger
 from requests.exceptions import HTTPError
-from urllib.parse import urlparse
-from decimal import Decimal
 
 from bson import ObjectId
 
-from django.conf import settings
 from django.http import HttpResponse
 
 from wstore.asset_manager.resource_plugins.decorators import on_product_acquired
@@ -45,6 +41,7 @@ from wstore.ordering.ordering_management import OrderingManager
 from wstore.store_commons.database import get_database_connection
 from wstore.store_commons.resource import Resource
 from wstore.store_commons.utils.http import authentication_required, build_response, supported_request_mime_types
+from wstore.charging_engine.pricing_engine import PriceEngine
 
 logger = getLogger("wstore.default_logger")
 
@@ -319,61 +316,6 @@ class PaymentRefund(Resource):
 
 class PaymentPreview(Resource):
 
-    def _download_pricing(self, pop_id):
-        catalog = urlparse(settings.CATALOG)
-        price_url = "{}://{}{}/{}".format(
-            catalog.scheme, catalog.netloc, catalog.path + "/productOfferingPrice", pop_id
-        )
-
-        request = requests.get(price_url, verify=settings.VERIFY_REQUESTS)
-        pricing = request.json()
-        return pricing
-
-    def _process_price_component(self, component, options, aggregated):
-        # Check if the component needs to be applied
-        tail_value = None
-        if "prodSpecCharValueUse" in component:
-            conditions = {}
-
-            for val in component["prodSpecCharValueUse"]:
-                value = 'tailored'
-                if "productSpecCharacteristicValue" in val and len(val["productSpecCharacteristicValue"]) > 0:
-                    value = val["productSpecCharacteristicValue"][0]["value"]
-                conditions[val["name"].lower()] = value
-
-            found = 0
-            for option in options:
-                if option["name"].lower() in conditions:
-                    if conditions[option["name"].lower()] == 'tailored':
-                        found += 1
-                        tail_value = Decimal(option["value"])
-                        continue
-
-                    if str(conditions[option["name"].lower()]) == str(option["value"]):
-                        found += 1
-    
-            if len(conditions) != found:
-                # The component is not processed
-                return
-
-        if component["priceType"] not in aggregated:
-            aggregated[component["priceType"]] = {}
-
-        period_key = "onetime"
-        if "recurringChargePeriodType" in component and "recurringChargePeriodLength" in component:
-            period_key = "{} {}".format(component["recurringChargePeriodLength"], component["recurringChargePeriodType"])
-
-        if period_key not in aggregated[component["priceType"]]:
-            aggregated[component["priceType"]][period_key] = {
-                "value": Decimal('0')
-            }
-
-        if tail_value is None:
-            aggregated[component["priceType"]][period_key]["value"] += Decimal(component["price"]["value"])
-        else:
-            tailored_price = Decimal(component["price"]["value"]) * tail_value
-            aggregated[component["priceType"]][period_key]["value"] += tailored_price
-
     @supported_request_mime_types(("application/json",))
     #@authentication_required
     def create(self, request):
@@ -381,57 +323,16 @@ class PaymentPreview(Resource):
             "orderTotalPrice": []
         }
 
-        aggregated = {}
         try:
             data = json.loads(request.body)
 
-            item = data["productOrderItem"][0]
-            # 1) Download the POP
-            pop_id = item["itemTotalPrice"][0]["productOfferingPrice"]["id"]
-            
-            pricing = self._download_pricing(pop_id)
-
-            # If the price is a bundle download the components
-            to_process = []
-            if pricing["isBundle"]:
-                to_process = [self._download_pricing(pop["id"]) for pop in pricing["bundledPopRelationship"]]
-            else:
-                to_process = [pricing]
-
-            options = []
-            if "product" in item and "productCharacteristic" in item["product"]:
-                options = item["product"]["productCharacteristic"]
-
-            for component in to_process:
-                self._process_price_component(component, options, aggregated)
-
+            price_engine = PriceEngine()
+            response = {
+            "orderTotalPrice": price_engine.calculate_prices(data)
+        }
         except:
             return build_response(request, 400, "Invalid order format")
-
-        # If the POP is not a bundle check the pricing
-        # If the bundle is a pop download the models
-
-        # If a characteristic has been defined check if the component have to be applied
-        # If the charactristic is tailored apply the value
-
-        for priceType in aggregated.keys():
-            for period in aggregated[priceType].keys():
-                response["orderTotalPrice"].append({
-                    "priceType": priceType,
-                    "recurringChargePeriod": period,
-                    "price": {
-                        "taxRate": 0,
-                        "dutyFreeAmount": {
-                            "unit": "EUR",
-                            "value": str(aggregated[priceType][period]['value'])
-                        },
-                        "taxIncludedAmount": {
-                            "unit": "EUR",
-                            "value": str(aggregated[priceType][period]['value'])
-                        }
-                    },
-                    "priceAlteration": []
-                })
+        
 
         # Check if a discount needs to be applied
         return HttpResponse(
